aadfs

+++ src/Window.zig
@@ -11,8 +11,8 @@ inner: switch (lib.platform) {
     .X11 => x11.Context,
     .Windows => unreachable,
 },
-width: u32,
-height: u32,
+width: u16,
+height: u16,
 /// Set to true when you want to close the window.
 should_close: bool = false,
